Split expectation and variance into two kernels
As neither the expectation values or variance depend on the other, both
can be calculated simultaneously. Using different kernels should allow
the calculation of both to begin as soon as possible.

This is however less space efficient than a single kernel as it requires
two pipes for the same data (one for each of the expectation and
variance kernels)