Reset cholesky solve results iteration
It seems that declaring these variables in the loop in not enough to
initialise them to zero.

Note that, if the forward and back substitution steps did not have their
innermost loops unrolled they could iterate from [0,i) instead of
[0,NX). This would mean one element of x, y could be set to zero each
iteration. It is unclear whether this would be more performant than
allowing the innermost loops to be unrolled, and how either approach
would scale with problem size.