Update demos to use new interfaces
This commit updates several of the demos to work with the new
interface to the GP class. In particular, priors and mean functions
had to be changed, and some additional calls to the validation
methods have been used to replace existing validation code.

Note that the MICE code does not currently work with the new
interface, as this will require some re-working of the code.
This demo has been left unchanged.