Fix GP Demos (#236)
* Fix GP Demos

Bugfixes to correct the GP demos to ensure that they run with the new
implementation. Changes include

- Update Patsy integration to allow using formulae with LHS

* Update demos to use new interfaces

This commit updates several of the demos to work with the new
interface to the GP class. In particular, priors and mean functions
had to be changed, and some additional calls to the validation
methods have been used to replace existing validation code.

Note that the MICE code does not currently work with the new
interface, as this will require some re-working of the code.
This demo has been left unchanged.

* Fix to code to compute design matrices

Fixes a bug that arises when a LHS is provided for a formula. Corrected
the code and added a unit test that catches this situation.

* Fix to R demo for GPs

Corrected the gp_demo.R file to correctly use the new prior
interface. Verified that the demo runs correctly and can be
used as an example script.

* Remove WIP demo tests from index

Test file to run all demos doesn't currently work due to issues with
RNG and plot displays, so I am removing from the index for the time
being. This issue will be addressed at another time.

* test GPU gp initialization with patsy formula in a couple of formats

* fix typo in last commit

* Increment version number for merge

* Fixed typo in version number bump

Co-authored-by: nbarlowATI <nbarlow@turing.ac.uk>